@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 0.3.3+303
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,26 +27,22 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  tfl_api_client:
-    git:
-      url: https://github.com/tnc1997/dart-tfl-api-client.git
-      ref: tfl-api-client_v0.4.0
-
-  cupertino_icons: ^0.1.3
-  http: ^0.12.1
-  intl: ^0.16.1
-  mdi: ^3.0.0
-  package_info: ^0.4.1
-  provider: ^4.1.3
-  shared_preferences: ^0.5.7
-  url_launcher: ^5.4.10
+  http: ^0.12.0
+  intl: ^0.17.0-nullsafety
+  mdi: ^5.0.0-nullsafety
+  package_info: ^0.4.0
+  provider: ^4.0.0
+  shared_preferences: ^0.5.0
+  tfl_api_client: ^1.0.0-nullsafety
+  url_launcher: ^5.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  mockito: ^4.1.1
-  pedantic: ^1.9.0
+  build_runner: ^1.0.0
+  mockito: ^5.0.0-nullsafety
+  pedantic: ^1.10.0-nullsafety
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
