@@ -3,15 +3,14 @@ import 'package:mdi/mdi.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 import 'package:tfl_api_explorer/src/specifications/road_display_name_specification.dart';
 import 'package:tfl_api_explorer/src/widgets/circular_progress_indicator_future_builder.dart';
-import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
 import 'package:tfl_api_explorer/src/widgets/road_list_tile.dart';
 
-class RoadSearchDelegate extends SearchDelegate<Road> {
+class RoadSearchDelegate extends SearchDelegate<RoadCorridor> {
   RoadSearchDelegate({
-    @required this.roadsFuture,
+    required this.roadsFuture,
   });
 
-  final Future<List<Road>> roadsFuture;
+  final Future<List<RoadCorridor>> roadsFuture;
 
   @override
   ThemeData appBarTheme(BuildContext context) {
@@ -37,64 +36,72 @@ class RoadSearchDelegate extends SearchDelegate<Road> {
     return IconButton(
       icon: Icon(Mdi.arrowLeft),
       onPressed: () {
-        close(context, null);
+        close(context, null as RoadCorridor);
       },
     );
   }
 
   @override
   Widget buildResults(BuildContext context) {
-    return CircularProgressIndicatorFutureBuilder<List<Road>>(
+    return CircularProgressIndicatorFutureBuilder<List<RoadCorridor>>(
       future: roadsFuture,
       builder: (context, data) {
         final specification = RoadDisplayNameSpecification(
           displayName: query,
         );
 
-        final roads = data.where(specification.isSatisfiedBy).toList();
+        final roads = data?.where(specification.isSatisfiedBy).toList();
 
-        return ListView.builder(
-          itemBuilder: (context, index) {
-            return RoadListTile(
-              road: roads[index],
-              onTap: () {
-                close(context, roads[index]);
-              },
-            );
-          },
-          itemCount: roads.length,
-        );
+        if (roads != null) {
+          return ListView.builder(
+            itemBuilder: (context, index) {
+              return RoadListTile(
+                road: roads[index],
+                onTap: () {
+                  close(context, roads[index]);
+                },
+              );
+            },
+            itemCount: roads.length,
+          );
+        } else {
+          return Container();
+        }
       },
     );
   }
 
   @override
   Widget buildSuggestions(BuildContext context) {
-    return CircularProgressIndicatorFutureBuilder<List<Road>>(
+    return CircularProgressIndicatorFutureBuilder<List<RoadCorridor>>(
       future: roadsFuture,
       builder: (context, data) {
         final specification = RoadDisplayNameSpecification(
           displayName: query,
         );
 
-        final roads = data.where(specification.isSatisfiedBy).toList();
+        final roads = data?.where(specification.isSatisfiedBy).toList();
 
-        return ListView.builder(
-          itemBuilder: (context, index) {
-            return ListTile(
-              title: NullableText(
-                roads[index].displayName,
-                overflow: TextOverflow.ellipsis,
-              ),
-              onTap: () {
-                query = roads[index].displayName;
+        if (roads != null) {
+          return ListView.builder(
+            itemBuilder: (context, index) {
+              return ListTile(
+                title: Text(
+                  roads[index].displayName ?? 'Unknown',
+                  overflow: TextOverflow.ellipsis,
+                ),
+                onTap: () {
+                  query = roads[index].displayName ?? '';
 
-                showResults(context);
-              },
-            );
-          },
-          itemCount: roads.length,
-        );
+                  showResults(context);
+                },
+              );
+            },
+            itemCount: roads.length,
+          );
+        } else {
+          return Container();
+        }
       },
     );
   }
