@@ -3,34 +3,25 @@ import 'package:http/http.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 
 class AuthenticationChangeNotifier extends ChangeNotifier {
-  String _appId;
+  String? _appKey;
 
-  String _appKey;
+  Client? _client;
 
-  Client _client;
+  String? get appKey => _appKey;
 
-  String get appId => _appId;
+  Client? get client => _client;
 
-  String get appKey => _appKey;
-
-  Client get client => _client;
-
-  void login(String appId, String appKey) {
-    if (appId == null) throw ArgumentError.notNull('appId');
-    if (appKey == null) throw ArgumentError.notNull('appKey');
-
-    _appId = appId;
+  void login(String appKey) {
     _appKey = appKey;
 
     _client?.close();
 
-    _client = clientViaAppIdAppKey(_appId, _appKey);
+    _client = clientViaAppKey(appKey);
 
     notifyListeners();
   }
 
   void logout() {
-    _appId = null;
     _appKey = null;
 
     _client?.close();
