@@ -1,14 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 import 'package:tfl_api_explorer/src/pages/car_parks/car_park_bays_page.dart';
-import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
 
 class CarParkPage extends StatelessWidget {
   static const routeName = '/car_parks/:id';
 
   CarParkPage({
-    Key key,
-    @required this.carPark,
+    Key? key,
+    required this.carPark,
   }) : super(
           key: key,
         );
@@ -19,14 +18,14 @@ class CarParkPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(carPark.id),
+        title: Text(carPark.id ?? 'Unknown'),
       ),
       body: ListView(
         children: [
           ListTile(
             title: Text('Name'),
-            subtitle: NullableText(
-              carPark.commonName,
+            subtitle: Text(
+              carPark.commonName ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
