@@ -1,65 +1,65 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
-import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
 
 class LineRoutePage extends StatelessWidget {
   static const routeName = '/line_routes/:id';
 
   LineRoutePage({
-    Key key,
-    @required this.lineRoute,
+    Key? key,
+    required this.lineRoute,
   }) : super(
           key: key,
         );
 
-  final LineRoute lineRoute;
+  final MatchedRoute lineRoute;
 
   @override
   Widget build(BuildContext context) {
+    final validTo = lineRoute.validTo;
+    final validFrom = lineRoute.validFrom;
+
     return Scaffold(
       appBar: AppBar(
-        title: Text(lineRoute.name),
+        title: Text(lineRoute.name ?? 'Unknown'),
       ),
       body: ListView(
         children: <Widget>[
           ListTile(
             title: Text('Origination'),
-            subtitle: NullableText(
-              lineRoute.originationName,
+            subtitle: Text(
+              lineRoute.originationName ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('Destination'),
-            subtitle: NullableText(
-              lineRoute.destinationName,
+            subtitle: Text(
+              lineRoute.destinationName ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('Service type'),
-            subtitle: NullableText(
-              lineRoute.serviceType,
+            subtitle: Text(
+              lineRoute.serviceType ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
-          ListTile(
-            title: Text('Valid to'),
-            subtitle: Text(
-              DateFormat.yMMMd().format(
-                lineRoute.validTo,
+          if (validTo != null)
+            ListTile(
+              title: Text('Valid to'),
+              subtitle: Text(
+                DateFormat.yMMMd().format(validTo),
               ),
             ),
-          ),
-          ListTile(
-            title: Text('Valid from'),
-            subtitle: Text(
-              DateFormat.yMMMd().format(
-                lineRoute.validFrom,
+          if (validFrom != null)
+            ListTile(
+              title: Text('Valid from'),
+              subtitle: Text(
+                DateFormat.yMMMd().format(validFrom),
               ),
             ),
-          ),
         ],
       ),
     );
