@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
-import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
 
 class LineStatusPage extends StatelessWidget {
   static const routeName = '/line_statuses/:id';
 
   LineStatusPage({
-    Key key,
-    @required this.lineStatus,
+    Key? key,
+    required this.lineStatus,
   }) : super(
           key: key,
         );
@@ -18,12 +17,12 @@ class LineStatusPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(lineStatus.statusSeverityDescription),
+        title: Text(lineStatus.statusSeverityDescription ?? 'Unknown'),
       ),
       body: Padding(
         padding: const EdgeInsets.all(16.0),
-        child: NullableText(
-          lineStatus.reason,
+        child: Text(
+          lineStatus.reason ?? 'Unknown',
           textAlign: TextAlign.justify,
         ),
       ),
