@@ -1,61 +1,60 @@
 import 'package:flutter/material.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 import 'package:tfl_api_explorer/src/pages/roads/road_road_disruptions_page.dart';
-import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class RoadPage extends StatelessWidget {
   static const routeName = '/roads/:id';
 
   RoadPage({
-    Key key,
-    @required this.road,
+    Key? key,
+    required this.road,
   }) : super(
           key: key,
         );
 
-  final Road road;
+  final RoadCorridor road;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(road.id),
+        title: Text(road.id ?? 'Unknown'),
       ),
       body: ListView(
         children: [
           ListTile(
             title: Text('Name'),
-            subtitle: NullableText(
-              road.displayName,
+            subtitle: Text(
+              road.displayName ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('Group'),
-            subtitle: NullableText(
-              road.group,
+            subtitle: Text(
+              road.group ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('Status severity'),
-            subtitle: NullableText(
-              road.statusSeverity,
+            subtitle: Text(
+              road.statusSeverity ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('Bounds'),
-            subtitle: NullableText(
-              road.bounds,
+            subtitle: Text(
+              road.bounds ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('URL'),
-            subtitle: NullableText(
-              road.url,
+            subtitle: Text(
+              road.url ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
             onTap: road.url != null
