@@ -1,14 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 import 'package:tfl_api_explorer/src/pages/stop_point_sequences/stop_point_sequence_stop_points_page.dart';
-import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
 
 class StopPointSequencePage extends StatelessWidget {
   static const routeName = '/stop_point_sequences/:id';
 
   StopPointSequencePage({
-    Key key,
-    @required this.stopPointSequence,
+    Key? key,
+    required this.stopPointSequence,
   }) : super(
           key: key,
         );
@@ -25,15 +24,15 @@ class StopPointSequencePage extends StatelessWidget {
         children: <Widget>[
           ListTile(
             title: Text('Line name'),
-            subtitle: NullableText(
-              stopPointSequence.lineName,
+            subtitle: Text(
+              stopPointSequence.lineName ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('Service type'),
-            subtitle: NullableText(
-              stopPointSequence.serviceType,
+            subtitle: Text(
+              stopPointSequence.serviceType ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
