@@ -3,15 +3,14 @@ import 'package:tfl_api_client/tfl_api_client.dart';
 import 'package:tfl_api_explorer/src/pages/stop_points/stop_point_additional_properties_page.dart';
 import 'package:tfl_api_explorer/src/pages/stop_points/stop_point_lines_page.dart';
 import 'package:tfl_api_explorer/src/pages/stop_points/stop_point_modes_page.dart';
-import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class StopPointPage extends StatelessWidget {
   static const routeName = '/stop_points/:id';
 
   StopPointPage({
-    Key key,
-    @required this.stopPoint,
+    Key? key,
+    required this.stopPoint,
   }) : super(
           key: key,
         );
@@ -22,56 +21,56 @@ class StopPointPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(stopPoint.id),
+        title: Text(stopPoint.id ?? 'Unknown'),
       ),
       body: ListView(
         children: <Widget>[
           ListTile(
             title: Text('Name'),
-            subtitle: NullableText(
-              stopPoint.commonName,
+            subtitle: Text(
+              stopPoint.commonName ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('ICS code'),
-            subtitle: NullableText(
-              stopPoint.icsCode,
+            subtitle: Text(
+              stopPoint.icsCode ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('SMS code'),
-            subtitle: NullableText(
-              stopPoint.smsCode,
+            subtitle: Text(
+              stopPoint.sMSCode ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('Stop type'),
-            subtitle: NullableText(
-              stopPoint.stopType,
+            subtitle: Text(
+              stopPoint.stopType ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('Accessibility'),
-            subtitle: NullableText(
-              stopPoint.accessibilitySummary,
+            subtitle: Text(
+              stopPoint.accessibilitySummary ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('Hub NaPTAN code'),
-            subtitle: NullableText(
-              stopPoint.hubNaptanCode,
+            subtitle: Text(
+              stopPoint.hubNaptanCode ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('URL'),
-            subtitle: NullableText(
-              stopPoint.url,
+            subtitle: Text(
+              stopPoint.url ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
             onTap: stopPoint.url != null
@@ -84,22 +83,22 @@ class StopPointPage extends StatelessWidget {
           ),
           ListTile(
             title: Text('Place type'),
-            subtitle: NullableText(
-              stopPoint.placeType,
+            subtitle: Text(
+              stopPoint.placeType ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('Lat'),
-            subtitle: NullableText(
-              stopPoint.lat?.toString(),
+            subtitle: Text(
+              stopPoint.lat?.toString() ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
           ListTile(
             title: Text('Lon'),
-            subtitle: NullableText(
-              stopPoint.lon?.toString(),
+            subtitle: Text(
+              stopPoint.lon?.toString() ?? 'Unknown',
               overflow: TextOverflow.ellipsis,
             ),
           ),
