@@ -1,10 +1,9 @@
-import 'package:flutter/foundation.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 import 'package:tfl_api_explorer/src/specifications/specification.dart';
 
 class LineModeNameSpecification extends Specification<Line> {
   LineModeNameSpecification({
-    @required this.modeName,
+    required this.modeName,
   });
 
   final String modeName;
