@@ -1,17 +1,16 @@
-import 'package:flutter/foundation.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 import 'package:tfl_api_explorer/src/specifications/specification.dart';
 
 class StopPointModesSpecification extends Specification<StopPoint> {
   StopPointModesSpecification({
-    @required this.modes,
+    required this.modes,
   });
 
   final Set<String> modes;
 
   @override
   bool isSatisfiedBy(StopPoint value) {
-    return value.modes.toSet().containsAll(modes);
+    return value.modes?.toSet().containsAll(modes) ?? false;
   }
 
   @override
