@@ -1,11 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
-import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
 
 class BayListTile extends StatelessWidget {
   BayListTile({
-    Key key,
-    @required this.bay,
+    Key? key,
+    required this.bay,
   }) : super(
           key: key,
         );
@@ -14,19 +13,26 @@ class BayListTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final occupied = bay.occupied;
+    final bayCount = bay.bayCount;
+
     return ListTile(
-      title: NullableText(
-        bay.bayType,
+      title: Text(
+        bay.bayType ?? 'Unknown',
         overflow: TextOverflow.ellipsis,
       ),
-      subtitle: NullableText(
-        '${bay.occupied} / ${bay.bayCount}',
-        overflow: TextOverflow.ellipsis,
-      ),
-      trailing: CircularProgressIndicator(
-        value: bay.occupied / bay.bayCount,
-        backgroundColor: Theme.of(context).backgroundColor,
-      ),
+      subtitle: occupied != null && bayCount != null
+          ? Text(
+              '$occupied / $bayCount',
+              overflow: TextOverflow.ellipsis,
+            )
+          : null,
+      trailing: occupied != null && bayCount != null
+          ? CircularProgressIndicator(
+              value: occupied / bayCount,
+              backgroundColor: Theme.of(context).backgroundColor,
+            )
+          : null,
     );
   }
 }
