@@ -3,30 +3,26 @@ import 'package:url_launcher/url_launcher.dart';
 
 class LoginForm extends StatefulWidget {
   const LoginForm({
-    Key key,
+    Key? key,
     this.formKey,
-    this.appIdController,
     this.appKeyController,
     this.onSubmitted,
   }) : super(
           key: key,
         );
 
-  final GlobalKey<FormState> formKey;
+  final GlobalKey<FormState>? formKey;
 
-  final TextEditingController appIdController;
+  final TextEditingController? appKeyController;
 
-  final TextEditingController appKeyController;
-
-  final void Function() onSubmitted;
+  final void Function()? onSubmitted;
 
   @override
   _LoginFormState createState() => _LoginFormState();
 }
 
 class _LoginFormState extends State<LoginForm> {
-  FocusNode _appIdFocusNode;
-  FocusNode _appKeyFocusNode;
+  late FocusNode _appKeyFocusNode;
 
   @override
   Widget build(BuildContext context) {
@@ -39,22 +35,6 @@ class _LoginFormState extends State<LoginForm> {
               bottom: 8.0,
             ),
             child: TextFormField(
-              controller: widget.appIdController,
-              focusNode: _appIdFocusNode,
-              decoration: InputDecoration(
-                labelText: 'App ID',
-              ),
-              onFieldSubmitted: (value) {
-                _appKeyFocusNode.requestFocus();
-              },
-              validator: _validator,
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(
-              vertical: 8.0,
-            ),
-            child: TextFormField(
               controller: widget.appKeyController,
               focusNode: _appKeyFocusNode,
               decoration: InputDecoration(
@@ -63,8 +43,9 @@ class _LoginFormState extends State<LoginForm> {
               onFieldSubmitted: (value) {
                 _appKeyFocusNode.unfocus();
 
-                if (widget.onSubmitted != null) {
-                  widget.onSubmitted();
+                final onSubmitted = widget.onSubmitted;
+                if (onSubmitted != null) {
+                  onSubmitted();
                 }
               },
               validator: _validator,
@@ -91,7 +72,6 @@ class _LoginFormState extends State<LoginForm> {
 
   @override
   void dispose() {
-    _appIdFocusNode.dispose();
     _appKeyFocusNode.dispose();
 
     super.dispose();
@@ -101,12 +81,11 @@ class _LoginFormState extends State<LoginForm> {
   void initState() {
     super.initState();
 
-    _appIdFocusNode = FocusNode();
     _appKeyFocusNode = FocusNode();
   }
 
-  String _validator(String value) {
-    if (value.isEmpty) {
+  String? _validator(String? value) {
+    if (value?.isEmpty ?? true) {
       return 'Please enter a value.';
     }
 
