@@ -2,12 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 import 'package:tfl_api_explorer/src/pages/predictions/prediction_page.dart';
-import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
 
 class PredictionListTile extends StatelessWidget {
   PredictionListTile({
-    Key key,
-    @required this.prediction,
+    Key? key,
+    required this.prediction,
   }) : super(
           key: key,
         );
@@ -16,15 +15,20 @@ class PredictionListTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final expectedArrival = prediction.expectedArrival;
+    final destinationName = prediction.destinationName;
+
     return ListTile(
-      title: NullableText(
-        prediction.id,
+      title: Text(
+        prediction.id ?? 'Unknown',
         overflow: TextOverflow.ellipsis,
       ),
-      subtitle: Text(
-        '${DateFormat.Hm().format(prediction.expectedArrival)} - ${prediction.destinationName}',
-        overflow: TextOverflow.ellipsis,
-      ),
+      subtitle: expectedArrival != null && destinationName != null
+          ? Text(
+              '${DateFormat.Hm().format(expectedArrival)} - $destinationName',
+              overflow: TextOverflow.ellipsis,
+            )
+          : null,
       onTap: () {
         Navigator.of(context).pushNamed(
           PredictionPage.routeName,
