@@ -1,29 +1,28 @@
 import 'package:flutter/material.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
-import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
 
 class RoadListTile extends StatelessWidget {
   RoadListTile({
-    Key key,
-    @required this.road,
+    Key? key,
+    required this.road,
     this.onTap,
   }) : super(
           key: key,
         );
 
-  final Road road;
+  final RoadCorridor road;
 
-  final void Function() onTap;
+  final void Function()? onTap;
 
   @override
   Widget build(BuildContext context) {
     return ListTile(
-      title: NullableText(
-        road.id,
+      title: Text(
+        road.id ?? 'Unknown',
         overflow: TextOverflow.ellipsis,
       ),
-      subtitle: NullableText(
-        road.displayName,
+      subtitle: Text(
+        road.displayName ?? 'Unknown',
         overflow: TextOverflow.ellipsis,
       ),
       onTap: onTap,
