@@ -10,7 +10,7 @@ import 'package:tfl_api_explorer/src/pages/stop_points/stop_points_page.dart';
 
 class TflApiExplorerDrawer extends StatelessWidget {
   TflApiExplorerDrawer({
-    Key key,
+    Key? key,
   }) : super(
           key: key,
         );
