@@ -4,9 +4,8 @@ import 'package:provider/provider.dart';
 import 'package:tfl_api_client/tfl_api_client.dart';
 import 'package:tfl_api_explorer/src/notifiers/line_filters_change_notifier.dart';
 import 'package:tfl_api_explorer/src/widgets/circular_progress_indicator_future_builder.dart';
-import 'package:tfl_api_explorer/src/widgets/nullable_text.dart';
-import 'package:tfl_api_explorer/src/widgets/tfl_api_explorer_drawer.dart';
 import 'package:tfl_api_explorer/src/widgets/line_list_tile.dart';
+import 'package:tfl_api_explorer/src/widgets/tfl_api_explorer_drawer.dart';
 
 class LinesPage extends StatefulWidget {
   static const routeName = '/lines';
@@ -16,7 +15,7 @@ class LinesPage extends StatefulWidget {
 }
 
 class _LinesPageState extends State<LinesPage> {
-  Future<List<Line>> _linesFuture;
+  late Future<List<Line>> _linesFuture;
 
   @override
   Widget build(BuildContext context) {
@@ -45,16 +44,20 @@ class _LinesPageState extends State<LinesPage> {
         future: _linesFuture,
         builder: (context, data) {
           final lines =
-              data.where(lineFiltersChangeNotifier.areSatisfiedBy).toList();
+              data?.where(lineFiltersChangeNotifier.areSatisfiedBy).toList();
 
-          return ListView.builder(
-            itemBuilder: (context, index) {
-              return LineListTile(
-                line: lines[index],
-              );
-            },
-            itemCount: lines.length,
-          );
+          if (lines != null) {
+            return ListView.builder(
+              itemBuilder: (context, index) {
+                return LineListTile(
+                  line: lines[index],
+                );
+              },
+              itemCount: lines.length,
+            );
+          } else {
+            return Container();
+          }
         },
       ),
       drawer: TflApiExplorerDrawer(),
@@ -65,7 +68,10 @@ class _LinesPageState extends State<LinesPage> {
   void initState() {
     super.initState();
 
-    _linesFuture = context.read<TflApi>().lines.get();
+    _linesFuture = context
+        .read<TflApiClient>()
+        .lines
+        .getByModeByPathModes(['bus', 'tube']);
   }
 }
 
@@ -75,7 +81,7 @@ class _LineFiltersPage extends StatefulWidget {
 }
 
 class _LineFiltersPageState extends State<_LineFiltersPage> {
-  Future<List<Mode>> _lineModesFuture;
+  late Future<List<Mode>> _lineModesFuture;
 
   @override
   Widget build(BuildContext context) {
@@ -97,26 +103,30 @@ class _LineFiltersPageState extends State<_LineFiltersPage> {
       body: CircularProgressIndicatorFutureBuilder<List>(
         future: Future.wait([_lineModesFuture]),
         builder: (context, data) {
-          return ListView(
-            children: <Widget>[
-              ExpansionTile(
-                title: Text('Mode name'),
-                children: (data[0] as List<Mode>).map((mode) {
-                  return RadioListTile<String>(
-                    value: mode.modeName,
-                    groupValue: lineFiltersChangeNotifier.modeName,
-                    onChanged: (value) {
-                      lineFiltersChangeNotifier.modeName = value;
-                    },
-                    title: NullableText(
-                      mode.modeName,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                  );
-                }).toList(),
-              ),
-            ],
-          );
+          if (data != null) {
+            return ListView(
+              children: <Widget>[
+                ExpansionTile(
+                  title: Text('Mode name'),
+                  children: (data[0] as List<Mode>).map((mode) {
+                    return RadioListTile<String>(
+                      value: mode.modeName!,
+                      groupValue: lineFiltersChangeNotifier.modeName,
+                      onChanged: (value) {
+                        lineFiltersChangeNotifier.modeName = value;
+                      },
+                      title: Text(
+                        mode.modeName ?? 'Unknown',
+                        overflow: TextOverflow.ellipsis,
+                      ),
+                    );
+                  }).toList(),
+                ),
+              ],
+            );
+          } else {
+            return Container();
+          }
         },
       ),
     );
@@ -126,6 +136,6 @@ class _LineFiltersPageState extends State<_LineFiltersPage> {
   void initState() {
     super.initState();
 
-    _lineModesFuture = context.read<TflApi>().lineModes.get();
+    _lineModesFuture = context.read<TflApiClient>().lines.metaModes();
   }
 }
